Motivation:
A microbenchmark will be useful to get a baseline for performance.

Modifications:
Introduce a new microbenchmark which tests the Http2DefaultFrameWriter.

Result:
Microbenchmark exists to test performance.