@Scottmitch 2015-02-04
@nmittler @louiscryan - FYI.  I originally was creating a local client/server connection to do the writing.  However memory became a problem very quickly (possibly because I wasn't flushing enough, and buffers weren't released)?  Anyhow I isolated the benchmarck from transport type (NIO/OIO/EPoll, etc..) for the time being.  Seems like this is an important piece to the puzzle.  Let me know what you think.