@Scottmitch 2015-02-04
I have added the transport sensitive benchmark.  I am just testing NIO currently.  My machine is under-powered but the benchmark is taking a good deal of time.  I can scale back what is being tested to save time if necessary.