@Scottmitch 2015-02-04
@nmittler - Sorry I squashed and pushed again.  I agree that it would be nicer to have a meaningful string displayed.  I will look into that.