@Scottmitch 2015-02-05
@louiscryan @nmittler - I have a first cut of Netty and JMH sharing an executor.  Please review.  After we get the code in an acceptable state we can run some benchmarks to get a baseline.