@Scottmitch 2015-02-05
I will also add `@Benchmark` methods for the other frame types (or at least headers) once we all approve of the general structure...