@louiscryan 2015-02-06
I think I need to understand the threading interactions here better to understand the design choices, my comments may or may not make sense if I have misunderstood something fundamental.

Some instructive comments would probably go along way