@louiscryan 2015-02-06
@normanmaurer 

I think we could use your wisdom here. The majority of work Scott has put into this PR is to avoid thread-hopping between the benchmark thread and the event-loop so we can isolate the performance of the HTTP2 frame writing code. 

He's had to do some pretty heavy lifting to get that working. Being able to avoid the thread hop seems like a generally useful feature for other benchmarks so we'd like it to be as broadly applicable as possible.