@Scottmitch 2015-02-06
+1. I was a bit hesitant to perturb the event loop and executor code just for benchmarking but there were a few areas that are worth discussion as @louiscryan has pointed out.