@Scottmitch 2015-03-06
@louiscryan @nmittler - Feel free to take another look.  I added header frames to the benchmark.  The `inEventLoop` hack is still in there for now.  Need to discuss with @normanmaurer still.