@trustin 2015-03-10
> Other than the hack in SingleThreadEventExecutor which seems a little dangerous to check-in, LGTM. Maybe we should fork that executor for the moment to avoid breaking others?

That's also a good idea. Just don't forget to file an issue :-)