@Scottmitch 2015-03-10
@louiscryan @trustin - No worries...wasn't planning on pulling in the hack :)  I may have a better way to use the same executor...standby.