@Scottmitch 2015-03-13
@louiscryan - Feel free to have another look. I was able to have JMH use Netty's event loop instead of the other way around (much cleaner).  I'm in the process of getting access to the benchmark machines....so no results to report yet.