@Scottmitch 2015-03-13
I'm working on making the build/code cross platform (more than just linux)...but you get the idea.