@louiscryan 2015-03-14
Much better! Comments are only nits so LGTM