@Scottmitch 2015-03-14
@louiscryan - I addressed all your comments except the following:

>Could move picking out values from the environment into instance variables to avoid timing that

Could you be more explicit?  Are you suggesting getting rid of the `ENVIRONMENTS` map?