@louiscryan 2015-03-16
If you can get @Setup working then it should allow you to pick values out of the environment and assign them to instance variables outside the benchmark methods. Not suggestion getting rid of ENVIRONMENTS.