@Scottmitch 2015-03-16
@louiscryan - I made some updates.  Let me know if this is what you had in mind.