@Scottmitch 2015-03-17
@louiscryan @nmittler - Results of the benchmark are below courtesy of @normanmaurer .  The `EPOLL` results have been omitted from the `Old` section because it got complicated pulling/building the Alpha2.SNAPSHOT with the new Alpha3.SNAPSHOT code.  The results are not as good as I expected them to be and may be a regression (for transports which hit the kernel).  Would you mind collaborating these results?  Keep in mind this PR has some changes in the HTTP/2 codec that should be applied/built/installed before generating the "new" values.

# New
```bash
# Run complete. Total time: 00:58:52

Benchmark                                        (dataType) (environmentType) (padding)   Mode   Samples        Score  Score error    Units
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL   EMBEDDED_POOLED         0  thrpt        10  2716523.185    34735.192    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL   EMBEDDED_POOLED       255  thrpt        10  1927658.082   276953.911    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL EMBEDDED_UNPOOLED         0  thrpt        10  1500349.363     8554.119    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL EMBEDDED_UNPOOLED       255  thrpt        10  1344056.521     9988.242    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL        NIO_POOLED         0  thrpt        10   671573.628    73477.551    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL        NIO_POOLED       255  thrpt        10   339660.490   133596.486    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL      NIO_UNPOOLED         0  thrpt        10   353586.410     6601.992    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL      NIO_UNPOOLED       255  thrpt        10   328953.882     5820.789    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL      EPOLL_POOLED         0  thrpt        10   369398.578     6324.707    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL      EPOLL_POOLED       255  thrpt        10   473184.616     3102.763    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL    EPOLL_UNPOOLED         0  thrpt        10   383522.893    10997.003    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL    EPOLL_UNPOOLED       255  thrpt        10   289786.597     6787.835    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL        OIO_POOLED         0  thrpt        10   219227.896     5548.026    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL        OIO_POOLED       255  thrpt        10   165572.254     7362.360    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL      OIO_UNPOOLED         0  thrpt        10   509301.135    13237.187    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL      OIO_UNPOOLED       255  thrpt        10   359229.323     1622.492    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM   EMBEDDED_POOLED         0  thrpt        10  2486845.129    14602.529    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM   EMBEDDED_POOLED       255  thrpt        10   431486.773     2380.164    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM EMBEDDED_UNPOOLED         0  thrpt        10  1507967.160    12293.474    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM EMBEDDED_UNPOOLED       255  thrpt        10   433113.942     2481.666    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM        NIO_POOLED         0  thrpt        10   630340.099    40438.204    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM        NIO_POOLED       255  thrpt        10   183996.566     9355.941    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM      NIO_UNPOOLED         0  thrpt        10   574013.021   143678.210    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM      NIO_UNPOOLED       255  thrpt        10   205091.975    14020.788    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM      EPOLL_POOLED         0  thrpt        10   593739.824    10568.011    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM      EPOLL_POOLED       255  thrpt        10   260486.527     4984.486    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM    EPOLL_UNPOOLED         0  thrpt        10   330242.923     4421.687    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM    EPOLL_UNPOOLED       255  thrpt        10   250527.448     1475.405    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM        OIO_POOLED         0  thrpt        10   210346.018     2625.663    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM        OIO_POOLED       255  thrpt        10   229465.952     1121.374    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM      OIO_UNPOOLED         0  thrpt        10   425448.883    13363.652    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData          MEDIUM      OIO_UNPOOLED       255  thrpt        10   230070.793     4042.460    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE   EMBEDDED_POOLED         0  thrpt        10   417151.015     1793.086    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE   EMBEDDED_POOLED       255  thrpt        10   406218.816     2601.957    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE EMBEDDED_UNPOOLED         0  thrpt        10   414142.938     2617.098    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE EMBEDDED_UNPOOLED       255  thrpt        10   409120.072     1857.436    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE        NIO_POOLED         0  thrpt        10   179972.043     7249.518    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE        NIO_POOLED       255  thrpt        10   183752.174     8044.027    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE      NIO_UNPOOLED         0  thrpt        10   168984.045     8050.265    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE      NIO_UNPOOLED       255  thrpt        10   173226.923    18448.392    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE      EPOLL_POOLED         0  thrpt        10   252811.147     1331.786    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE      EPOLL_POOLED       255  thrpt        10   243709.305     1958.636    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE    EPOLL_UNPOOLED         0  thrpt        10   251001.603      878.108    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE    EPOLL_UNPOOLED       255  thrpt        10   252410.313     1329.234    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE        OIO_POOLED         0  thrpt        10   225401.577     2508.765    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE        OIO_POOLED       255  thrpt        10   226101.627     1282.967    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE      OIO_UNPOOLED         0  thrpt        10   230974.141     5081.345    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           LARGE      OIO_UNPOOLED       255  thrpt        10   232638.914     1281.715    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO   EMBEDDED_POOLED         0  thrpt        10   410857.527     1839.714    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO   EMBEDDED_POOLED       255  thrpt        10   404040.363     4130.772    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO EMBEDDED_UNPOOLED         0  thrpt        10   422110.877     2271.087    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO EMBEDDED_UNPOOLED       255  thrpt        10   416525.917     3175.402    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO        NIO_POOLED         0  thrpt        10   173824.759    19511.220    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO        NIO_POOLED       255  thrpt        10   179274.368     7617.086    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO      NIO_UNPOOLED         0  thrpt        10   137075.140    46913.952    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO      NIO_UNPOOLED       255  thrpt        10   170836.216     9953.441    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO      EPOLL_POOLED         0  thrpt        10   249660.241     2142.401    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO      EPOLL_POOLED       255  thrpt        10   246735.837     1592.923    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO    EPOLL_UNPOOLED         0  thrpt        10   255295.931     1131.464    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO    EPOLL_UNPOOLED       255  thrpt        10   254232.922     1507.097    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO        OIO_POOLED         0  thrpt        10   226976.470     3317.594    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO        OIO_POOLED       255  thrpt        10   232651.030     4535.205    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO      OIO_UNPOOLED         0  thrpt        10   235484.249     3448.784    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           JUMBO      OIO_UNPOOLED       255  thrpt        10   237622.885     2056.521    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL   EMBEDDED_POOLED         0  thrpt        10   654091.372     3067.846    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL   EMBEDDED_POOLED       255  thrpt        10   623122.215     4479.039    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL EMBEDDED_UNPOOLED         0  thrpt        10   420124.476     2722.059    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL EMBEDDED_UNPOOLED       255  thrpt        10   403845.750     1239.727    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL        NIO_POOLED         0  thrpt        10   163132.031    12987.010    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL        NIO_POOLED       255  thrpt        10   142219.294     5541.151    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL      NIO_UNPOOLED         0  thrpt        10   174398.485     8749.610    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL      NIO_UNPOOLED       255  thrpt        10   176570.851    30955.847    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL      EPOLL_POOLED         0  thrpt        10   194248.855     8293.080    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL      EPOLL_POOLED       255  thrpt        10   159287.923     2383.820    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL    EPOLL_UNPOOLED         0  thrpt        10   267839.835     2177.752    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL    EPOLL_UNPOOLED       255  thrpt        10   229014.249     5687.889    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL        OIO_POOLED         0  thrpt        10   220389.883     3201.607    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL        OIO_POOLED       255  thrpt        10   134515.041    27372.733    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL      OIO_UNPOOLED         0  thrpt        10   221639.648     4351.306    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        SMALL      OIO_UNPOOLED       255  thrpt        10   177792.087     2836.893    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM   EMBEDDED_POOLED         0  thrpt        10   184815.501     1050.470    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM   EMBEDDED_POOLED       255  thrpt        10   182924.991     1205.901    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM EMBEDDED_UNPOOLED         0  thrpt        10   160127.948     1070.962    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM EMBEDDED_UNPOOLED       255  thrpt        10   157426.488      870.507    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM        NIO_POOLED         0  thrpt        10   114848.472    19144.159    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM        NIO_POOLED       255  thrpt        10    91658.561    17458.794    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM      NIO_UNPOOLED         0  thrpt        10   102551.175     6261.578    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM      NIO_UNPOOLED       255  thrpt        10   114888.777     1407.884    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM      EPOLL_POOLED         0  thrpt        10   121773.961     1915.757    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM      EPOLL_POOLED       255  thrpt        10   100081.145     1115.221    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM    EPOLL_UNPOOLED         0  thrpt        10   108005.211     1726.823    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM    EPOLL_UNPOOLED       255  thrpt        10   111430.405     1311.916    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM        OIO_POOLED         0  thrpt        10   110213.552      873.972    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM        OIO_POOLED       255  thrpt        10   105302.188      930.137    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM      OIO_UNPOOLED         0  thrpt        10   108987.815      705.885    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders       MEDIUM      OIO_UNPOOLED       255  thrpt        10   104958.872      895.484    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE   EMBEDDED_POOLED         0  thrpt        10     4155.118     3240.869    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE   EMBEDDED_POOLED       255  thrpt        10     4454.895     3198.548    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE EMBEDDED_UNPOOLED         0  thrpt        10     4197.794     2915.903    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE EMBEDDED_UNPOOLED       255  thrpt        10     4654.475     2180.757    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE        NIO_POOLED         0  thrpt        10     4651.748     2435.007    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE        NIO_POOLED       255  thrpt        10     4627.146     2341.969    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE      NIO_UNPOOLED         0  thrpt        10     4669.226     2314.705    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE      NIO_UNPOOLED       255  thrpt        10     4670.397     2351.694    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE      EPOLL_POOLED         0  thrpt        10     5176.039        9.823    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE      EPOLL_POOLED       255  thrpt        10     4660.009     2409.543    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE    EPOLL_UNPOOLED         0  thrpt        10     4729.848     2438.110    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE    EPOLL_UNPOOLED       255  thrpt        10     4728.788     2417.805    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE        OIO_POOLED         0  thrpt        10     4688.512     2355.558    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE        OIO_POOLED       255  thrpt        10     4663.747     2301.451    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE      OIO_UNPOOLED         0  thrpt        10     4749.271     2300.203    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        LARGE      OIO_UNPOOLED       255  thrpt        10     4666.449     2447.368    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO   EMBEDDED_POOLED         0  thrpt        10      623.616      285.499    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO   EMBEDDED_POOLED       255  thrpt        10      622.516      263.895    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO EMBEDDED_UNPOOLED         0  thrpt        10      555.033      235.061    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO EMBEDDED_UNPOOLED       255  thrpt        10      543.376      272.683    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO        NIO_POOLED         0  thrpt        10      675.462        3.655    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO        NIO_POOLED       255  thrpt        10      610.221      304.139    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO      NIO_UNPOOLED         0  thrpt        10      678.827        1.350    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO      NIO_UNPOOLED       255  thrpt        10      608.751      309.704    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO      EPOLL_POOLED         0  thrpt        10      609.973      296.977    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO      EPOLL_POOLED       255  thrpt        10      613.562      301.330    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO    EPOLL_UNPOOLED         0  thrpt        10      611.544      298.152    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO    EPOLL_UNPOOLED       255  thrpt        10      674.706        4.279    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO        OIO_POOLED         0  thrpt        10      615.185      286.411    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO        OIO_POOLED       255  thrpt        10      673.037        3.206    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO      OIO_UNPOOLED         0  thrpt        10      615.794      284.911    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeHeaders        JUMBO      OIO_UNPOOLED       255  thrpt        10      616.037      280.111    ops/s
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3,535.447 sec - in io.netty.microbench.http2.Http2FrameWriterBenchmark
```

# Old
```bash
# Run complete. Total time: 00:56:56

Benchmark                                        (dataType) (environmentType) (padding)   Mode   Samples        Score  Score error    Units
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL   EMBEDDED_POOLED         0  thrpt        10  2405609.855    11999.400    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL   EMBEDDED_POOLED       255  thrpt        10  1524131.478     7349.395    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL EMBEDDED_UNPOOLED         0  thrpt        10  1318838.264    15866.839    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL EMBEDDED_UNPOOLED       255  thrpt        10   476516.153     2223.194    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL        NIO_POOLED         0  thrpt        10   857070.669   145536.779    ops/s
i.n.m.h.Http2FrameWriterBenchmark.writeData           SMALL        NIO_POOLED       255  thrpt        10   502120.379     7285.615    ops/s
…